Finish MCI clearance
